Move questionnaire API tests out of common file, there are too many
